IOS-233 Replace unusable library picker by reusing VC from welcome flow
The "new" NYPLLibrariesListVC was taken verbatim from NYPLWelcomeScreenViewController.swift.
